Update Yolo examples, merge them together into one experiment with multiple subfolders
