+++ gen2-yolo/device-decoding/json/yolov4-tiny.json
@@ -6,13 +6,13 @@
         "input_size": "416x416",
         "NN_specific_metadata" :
         {
-            "classes" : 1,
+            "classes" : 80,
             "coordinates" : 4,
             "anchors" : [10,14, 23,27, 37,58, 81,82, 135,169, 344,319],
             "anchor_masks" :
             {
-                "side32" : [1,2,3],
-                "side16" : [3,4,5]
+                "side26" : [1,2,3],
+                "side13" : [3,4,5]
             },
             "iou_threshold" : 0.5,
             "confidence_threshold" : 0.5
