+++ gen2-yolo/device-decoding/json/yolov4.json
@@ -11,9 +11,9 @@
             "anchors" : [12, 16, 19, 36, 40, 28, 36, 75, 76, 55, 72, 146, 142, 110, 192, 243, 459, 401],
             "anchor_masks" : 
             {
-                "side52" : [0,1,2],
-                "side26" : [3,4,5],
-                "side13" : [6,7,8]
+                "side76" : [0,1,2],
+                "side38" : [3,4,5],
+                "side19" : [6,7,8]
             },
             "iou_threshold" : 0.5,
             "confidence_threshold" : 0.5
