+++ gen2-yolo/yolop/requirements.txt
@@ -1,0 +1,6 @@
+opencv-python
+depthai>=2.10.0
+torch>=1.7.0
+torchvision>=0.8.1
+numpy>=1.18.5
+blobconverter
