+++ gen2-yolo/yolop/vids/empty.txt
@@ -1,0 +1,1 @@
+Use `python download.py` to download sample data.
