--- gen2-yolop/requirements.txt
@@ -1,6 +1,0 @@
-opencv-python
-depthai>=2.10.0.0
-torch>=1.7.0
-torchvision>=0.8.1
-numpy>=1.18.5
-blobconverter
