--- gen2-yolop/vids/empty.txt
@@ -1,1 +1,0 @@
-Use `python download.py` to download sample data.
